mcp-crew/requirements.txt
# MCP-Crew - Requisitos
# Dependências principais
flask==2.3.3
redis==5.0.1
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.0

# Suporte para processamento paralelo
asyncio==3.4.3
aiohttp==3.9.1
# concurrent-futures-extra removido - não disponível no PyPI

# Serialização e processamento de dados
numpy==1.26.2
pandas==2.1.3
msgpack==1.0.7
protobuf==4.24.4

# Integração com Crew AI
crewai==0.121.1

# Segurança
pyjwt==2.8.0
cryptography>=43.0.1

# Utilitários
tqdm==4.66.1
loguru==0.7.2

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
